+++ src/components/DocumentSelector/index.tsx
@@ -10,6 +10,7 @@ import { TranslateCallback } from '~types/locales'
 import PageTitle from '../PageTitle'
 import { LocaleContext } from '~locales'
 import { DocumentTypeConfig, DocumentTypes, PoaTypes } from '~types/steps'
+import { PoASupportedCountry } from '~types/api'
 
 const always = () => true
 
@@ -19,17 +20,20 @@ export type Props = {
   country?: string
   type: DocumentTypes | PoaTypes
   autoFocusOnInitialScreenTitle?: boolean
+  countryList?: PoASupportedCountry[]
 } & StepComponentBaseProps
 
 // The 'type' value of these options must match the API document types.
 // See https://documentation.onfido.com/#document-types
 export abstract class DocumentSelectorBase extends Component<Props> {
   private defaultOptions: DocumentOptionsType[] | undefined = undefined
+  protected country: string | undefined = undefined
 
   private getOptions = (
     translate: (key: string, options?: Record<string, unknown>) => string
   ) => {
-    const { documentTypes, country = 'GBR' } = this.props
+    const { documentTypes, country } = this.props
+    const countryCode = country || this.country || 'GBR'
 
     if (!this.defaultOptions) {
       this.defaultOptions = generateDefaultOptions(
@@ -39,7 +43,8 @@ export abstract class DocumentSelectorBase extends Component<Props> {
     }
 
     const defaultDocOptions = this.defaultOptions.filter(
-      ({ checkAvailableInCountry = always }) => checkAvailableInCountry(country)
+      ({ checkAvailableInCountry = always }) =>
+        checkAvailableInCountry(countryCode)
     )
     const checkAvailableType = isEmpty(documentTypes)
       ? always
@@ -71,6 +76,7 @@ export abstract class DocumentSelectorBase extends Component<Props> {
 
   render() {
     const { className, country } = this.props
+    const countryCode = country || this.country || 'GBR'
 
     return (
       <LocaleContext.Consumer>
@@ -82,7 +88,7 @@ export abstract class DocumentSelectorBase extends Component<Props> {
           const { translate } = injectedProps
 
           const title = translate(this.titleTranslationKey(), {
-            country: !country || country === 'GBR' ? 'UK' : '',
+            country: !countryCode || countryCode === 'GBR' ? 'UK' : '',
           })
 
           const subTitle = translate(this.subTitleTranslationKey())