+++ src/components/Router/StepComponentMap.ts
@@ -2,7 +2,6 @@ import type { ComponentType } from 'preact'
 import Welcome from '../Welcome'
 import UserConsent from '../UserConsent'
 
-import CountrySelector from '../CountrySelector'
 import ImageQualityGuide from '../Uploader/ImageQualityGuide'
 import SelfieIntro from '../Photo/SelfieIntro'
 import {
@@ -33,6 +32,7 @@ import { buildStepFinder, hasOnePreselectedDocument } from '~utils/steps'
 import { getCountryDataForDocumentType } from '~supported-documents'
 
 import type {
+  CountryData,
   ExtendedStepConfig,
   ExtendedStepTypes,
   FlowVariants,
@@ -48,6 +48,8 @@ import PoAClientIntro from '../ProofOfAddress/PoAIntro'
 import PoADocumentSelector from '../ProofOfAddress/PoADocumentSelect'
 import Guidance from '../ProofOfAddress/Guidance'
 import { SelectIdentityDocument } from '../Select/IdentityDocumentSelector'
+import DocumentCountrySelector from 'components/CountrySelector/DocumentCountrySelector'
+import PoACountrySelector from 'components/CountrySelector/PoACountrySelector'
 
 let LazyAuth: ComponentType<StepComponentProps>
 
@@ -68,9 +70,11 @@ export const buildComponentsList = ({
   steps,
   mobileFlow,
   deviceHasCameraSupport,
+  poaDocumentCountry,
 }: {
   flow: FlowVariants
   documentType: DocumentTypes | undefined
+  poaDocumentCountry?: CountryData | undefined
   steps: StepConfig[]
   mobileFlow?: boolean
   deviceHasCameraSupport?: boolean
@@ -80,6 +84,7 @@ export const buildComponentsList = ({
   return flow === 'captureSteps'
     ? buildComponentsFromSteps(
         buildCaptureStepComponents(
+          poaDocumentCountry,
           documentType,
           mobileFlow,
           steps,
@@ -114,6 +119,7 @@ const shouldUseCameraForDocumentCapture = (
 }
 
 const buildCaptureStepComponents = (
+  poaDocumentCountry: CountryData | undefined,
   documentType: DocumentTypes | undefined,
   mobileFlow: boolean | undefined,
   steps: StepConfig[],
@@ -157,7 +163,13 @@ const buildCaptureStepComponents = (
         firstCaptureStepType === 'document'
       ),
     ],
-    poa: [...buildPoaComponents(mobileFlow, firstCaptureStepType === 'poa')],
+    poa: [
+      ...buildPoaComponents(
+        poaDocumentCountry,
+        mobileFlow,
+        firstCaptureStepType === 'poa'
+      ),
+    ],
     complete,
   }
 }
@@ -244,7 +256,9 @@ const buildNonPassportPreCaptureComponents = (
   const prependDocumentSelector = hasOnePreselectedDocument
     ? []
     : [SelectIdentityDocument]
-  const prependCountrySelector = showCountrySelection ? [CountrySelector] : []
+  const prependCountrySelector = showCountrySelection
+    ? [DocumentCountrySelector]
+    : []
   // @ts-ignore
   // TODO: convert DocumentSelector to TS
   return [...prependDocumentSelector, ...prependCountrySelector]
@@ -372,11 +386,13 @@ const buildDocumentComponents = (
 }
 
 const buildPoaComponents = (
+  poaDocumentCountry: CountryData | undefined,
   mobileFlow: boolean | undefined,
   isFirstCaptureStepInFlow: boolean | undefined
 ): ComponentType<StepComponentProps>[] => {
   const preCaptureComponents = [
     PoAClientIntro,
+    PoACountrySelector,
     PoADocumentSelector,
     Guidance as ComponentType<StepComponentProps>,
   ]