+++ test/mock-server/jwt.ts
@@ -14,8 +14,13 @@ export const generateToken = async (context: any) => {
 
   const payload = {
     exp: getNumericDate(EXP_IN_SECONDS),
-    payload:
-      'TeING9+F5TB8waiN7l94SI9chwJppCYTaxTqqOdJzilmKSxh2jqsbaZ2BJoe\naR/0PhrfPUvWG2inSeLTC5M+Rg==\n',
+    payload: {
+      app: 'ebbad7e0-5f3c-4a6e-bf96-fa65c172d12c',
+      client_uuid: 'ce69f18f-8006-423d-afda-fcaff0de2154',
+      is_sandbox: false,
+      ref: '*://*/*',
+      sardine_session: '1207fcd6-e76b-463d-b84a-fbcd17cf6f54',
+    },
     uuid: 'iwokFVedG18',
     enterprise_features: {
       cobrand: true,