+++ test/webtest/src/test/java/com/onfido/qa/websdk/test/PercyIT.java
@@ -2,6 +2,7 @@ package com.onfido.qa.websdk.test;
 
 import com.onfido.qa.websdk.PoADocumentType;
 import com.onfido.qa.websdk.page.Complete;
+import com.onfido.qa.websdk.page.PoADocumentSelection;
 import com.onfido.qa.websdk.page.PoAIntro;
 import com.onfido.qa.websdk.page.Welcome;
 import org.testng.SkipException;
@@ -35,9 +36,11 @@ public class PercyIT extends WebSdkIT {
         var intro = onfido().withSteps("poa").init(PoAIntro.class);
         takePercySnapshot("PoAIntro");
 
-        intro.startVerification();
-        takePercySnapshot("PoADocumentSelection");
+        var countrySelect = intro.startVerification();
+        takePercySnapshot("PoACountrySelect");
 
+        countrySelect.select("United", PoADocumentSelection.class);
+        takePercySnapshot("PoADocumentSelection");
     }
 
     @Test(dataProvider = "poaDocumentTypes", groups = {"percy"})
@@ -50,6 +53,7 @@ public class PercyIT extends WebSdkIT {
 
         onfido().withSteps("poa").init(PoAIntro.class)
                 .startVerification()
+                .select("United", PoADocumentSelection.class)
                 .select(documentType);
 
         takePercySnapshot("PoAGuidance-" + documentType.canonicalName());