+++ test/webtest/src/test/java/com/onfido/qa/websdk/test/ProofOfAddressIT.java
@@ -12,6 +12,7 @@ import com.onfido.qa.websdk.page.CrossDeviceSubmit;
 import com.onfido.qa.websdk.page.DocumentUpload;
 import com.onfido.qa.websdk.page.IdDocumentSelector;
 import com.onfido.qa.websdk.page.ImageQualityGuide;
+import com.onfido.qa.websdk.page.PoADocumentSelection;
 import com.onfido.qa.websdk.page.PoAIntro;
 import com.onfido.qa.websdk.sdk.PoAStep;
 import com.onfido.qa.websdk.sdk.Raw;
@@ -62,23 +63,36 @@ public class ProofOfAddressIT extends WebSdkIT {
 
     }
 
+    private static class SimpleCountry {
+        public String code;
+        public String name;
+
+        public SimpleCountry(String name, String code) {
+            this.name = name;
+            this.code = code;
+        }
+    }
+
     @DataProvider
-    public static Object[][] countries() {
-        return new Object[][]{
-                {"GBR"},
-                {"GER"}
+    public static Object[] countries() {
+        return new SimpleCountry[]{
+                new SimpleCountry("United", "GBR"),
+                new SimpleCountry("Germany", "DEU")
         };
     }
 
     @Test(description = "verify that only available options for PoA are shown", dataProvider = "countries")
-    public void testVerifyThatOnlyAvailableOptionsForPoAAreShown(String countryCode) {
+    public void testVerifyThatOnlyAvailableOptionsForPoAAreShown(SimpleCountry country) {
 
         var availableDocumentTypes = Arrays.stream(values())
-                                           .filter(x -> x.availableInCountry(countryCode))
+                                           .filter(x -> x.availableInCountry(country.code))
                                            .collect(Collectors.toList());
 
-        var documentSelection = onfido().withSteps(new PoAStep().withCountry(countryCode)).init(PoAIntro.class)
-                                        .startVerification();
+        var documentSelection = onfido()
+                .withSteps(new PoAStep())
+                .init(PoAIntro.class)
+                .startVerification()
+                .select(country.name, PoADocumentSelection.class);
 
         var options = documentSelection.getOptions();
 
@@ -97,7 +111,7 @@ public class ProofOfAddressIT extends WebSdkIT {
 
         var intro = onfido().withSteps("poa", "complete").init(PoAIntro.class);
 
-        assertThat(intro.title()).isEqualTo("Let’s verify your UK address");
+        assertThat(intro.title()).isEqualTo("Let’s verify your address");
 
         verifyCopy(intro.requirementsHeader(), "poa_intro.subtitle");
         assertThat(intro.firstRequirement()).isEqualTo("Shows your current address");
@@ -106,7 +120,8 @@ public class ProofOfAddressIT extends WebSdkIT {
 
         verifyCopy(intro.startVerificationButtonText(), "poa_intro.button_primary");
 
-        var documentSelection = intro.startVerification();
+        var countrySelector = intro.startVerification();
+        var documentSelection = countrySelector.select("United", PoADocumentSelection.class);
 
         assertThat(documentSelection.title()).isEqualTo("Select a UK document");
         verifyCopy(documentSelection.subTitle(), "doc_select.subtitle_poa");
@@ -126,11 +141,13 @@ public class ProofOfAddressIT extends WebSdkIT {
         upload.upload(NATIONAL_IDENTITY_CARD_PDF).clickConfirmButton(Complete.class);
     }
 
+    /*
     @Test(description = "should skip country selection screen with a preselected driver's license document type on PoA flow", groups = {"percy"})
     public void testShouldSkipCountrySelectionScreenWithAPreselectedDriverSLicenseDocumentTypeOnPoAFlow() {
 
         var upload = onfido().withSteps("poa").init(PoAIntro.class)
                              .startVerification()
+                             .select("United", PoADocumentSelection.class)
                              .select(COUNCIL_TAX)
                              .clickContinue();
 
@@ -139,12 +156,14 @@ public class ProofOfAddressIT extends WebSdkIT {
         upload.upload(UK_DRIVING_LICENCE_PNG);
 
     }
+    */
 
     @Test(description = "should successfully complete cross device e2e flow with PoA document and selfie upload", groups = {"percy", "tabs"})
     public void testShouldSuccessfullyCompleteCrossDeviceE2EFlowWithPoADocumentAndSelfieUpload() {
 
         var crossDeviceLink = onfido().withSteps("poa", "complete").init(PoAIntro.class)
                                       .startVerification()
+                                      .select("United", PoADocumentSelection.class)
                                       .select(BANK_BUILDING_SOCIETY_STATEMENT)
                                       .clickContinue().switchToCrossDevice()
                                       .getSecureLink();
@@ -168,12 +187,13 @@ public class ProofOfAddressIT extends WebSdkIT {
 
     }
 
-    @Test(description = "should allow PoA together with Document")
-    public void testPoaTogetherWithDocument() throws JsonProcessingException {
+    @Test(description = "should allow PoA then Document")
+    public void testPoaFollowedByDocument() throws JsonProcessingException {
         onfido().withSteps("poa", "document", "complete")
                 .withOnComplete(new Raw("(data) => {window.onCompleteData = data}"))
                 .init(PoAIntro.class)
                 .startVerification()
+                .select("United", PoADocumentSelection.class)
                 .select(BANK_BUILDING_SOCIETY_STATEMENT)
                 .clickContinue()
                 .upload(NATIONAL_IDENTITY_CARD_PDF)
@@ -196,5 +216,28 @@ public class ProofOfAddressIT extends WebSdkIT {
 
         assertThat(poa.type).isEqualTo("passport");
         assertThat(documentFront.type).isEqualTo("passport");
+    }
+
+    @Test(description = "should allow Document (Passport) then PoA")
+    public void testPassportDocFollowedByPoa() throws JsonProcessingException {
+        onfido().withSteps("document", "poa", "complete")
+                .withOnComplete(new Raw("(data) => {window.onCompleteData = data}"))
+                .init(IdDocumentSelector.class)
+                .select(PASSPORT, DocumentUpload.class)
+                .clickUploadButton(ImageQualityGuide.class)
+                .upload(PASSPORT_JPG)
+                .clickConfirmButton(PoAIntro.class)
+                .startVerification()
+                .select("United", PoADocumentSelection.class)
+                .select(BANK_BUILDING_SOCIETY_STATEMENT)
+                .clickContinue()
+                .upload(NATIONAL_IDENTITY_CARD_PDF)
+                .clickConfirmButton(Complete.class);
+
+        var json = (String) driver().executeScript("return JSON.stringify(window.onCompleteData)");
+        var completeData = objectMapper.readValue(json, CompleteData.class);
+        var poa = completeData.poa;
+
+        assertThat(poa).isNotNull();
     }
 }