+++ test/webtest/src/test/java/com/onfido/qa/websdk/test/UserConsentIT.java
@@ -1,18 +1,25 @@
 package com.onfido.qa.websdk.test;
 
-import com.onfido.qa.websdk.page.IdDocumentSelector;
-import com.onfido.qa.websdk.page.UserConsent;
+import com.onfido.qa.webdriver.common.Page;
+import com.onfido.qa.websdk.page.*;
 import org.openqa.selenium.By;
 import org.testng.annotations.Test;
 
+import java.util.Map;
+
+import static com.onfido.qa.websdk.DocumentType.PASSPORT;
 import static org.assertj.core.api.Assertions.assertThat;
 
 public class UserConsentIT extends WebSdkIT {
 
     public static final String CONSENT_FRAME_TITLE = "Onfido's privacy statement and Terms of Service";
 
+    public <T extends Page> T init(Class<T> pageClass, Object... steps) {
+        return onfido().withSdkConfiguration(Map.of("sdk_features", Map.of("enable_require_applicant_consents", true))).withSteps(steps).init(pageClass);
+    }
+
     private UserConsent init() {
-        return onfido().withSteps("userConsent", "document").init(UserConsent.class);
+        return init(UserConsent.class,"document");
     }
 
     @Test(description = "should verify UI elements on the consent screen")
@@ -55,5 +62,31 @@ public class UserConsentIT extends WebSdkIT {
     public void testSecondaryClickOnDeclineModalUnmountsSdK() {
         init().declineUserConsent().clickSecondaryButton();
         assertThat(driver().findElements(By.id("#root > *"))).isEmpty();
+    }
+
+    @Test(description = "remove consents from navigation once accepted")
+    public void testRemoveConsentsFromNavigationOnceAccepted() {
+        var documentSelector = init().acceptUserConsent(IdDocumentSelector.class);
+
+        verifyCopy(documentSelector.title(), "doc_select.title");
+        assertThat(documentSelector.backArrow().isDisplayed()).isFalse();
+    }
+
+    @Test(description = "remove consents from history once accepted")
+    public void testRemoveConsentsFromHistoryOnceAccepted() {
+        var crossDevice = init(Welcome.class,"welcome", "document")
+                .continueToNextStep(UserConsent.class)
+                .acceptUserConsent(IdDocumentSelector.class)
+                .select(PASSPORT, DocumentUpload.class)
+                     .switchToCrossDevice();
+
+        assertThat(crossDevice.backArrow().isDisplayed()).isTrue();
+
+        var welcome = crossDevice
+                .back(DocumentUpload.class)
+                .back(IdDocumentSelector.class)
+                .back(Welcome.class);
+
+        verifyCopy(welcome.title(), "welcome.title");
     }
 }